# docs/swagger.yaml
basePath: /
definitions:
  ent.Todo:
    properties:
      description:
        description: Description holds the value of the "description" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      title:
        description: Title holds the value of the "title" field.
        type: string
    type: object
  ent.User:
    properties:
      id:
        description: ID of the ent.
        type: integer
      password:
        description: Password holds the value of the "password" field.
        type: string
      role:
        description: Role holds the value of the "role" field.
        type: string
      username:
        description: Username holds the value of the "username" field.
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a simple Todo App API.
  title: Todo App API
  version: "1.0"
paths:
  /todos:
    get:
      description: Bu endpoint, tüm todo'ları listeler.
      produces:
      - application/json
      responses:
        "200":
          description: Todo'lar başarıyla listelendi
          schema:
            items:
              $ref: '#/definitions/ent.Todo'
            type: array
        "500":
          description: Todo'lar listelenemedi
          schema:
            type: string
      summary: Get all todos
      tags:
      - todos
    post:
      consumes:
      - application/json
      description: Bu endpoint, yeni bir todo eklemek için kullanılır.
      parameters:
      - description: Todo nesnesi
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/ent.Todo'
      produces:
      - application/json
      responses:
        "201":
          description: Todo başarıyla eklendi
          schema:
            type: string
        "400":
          description: Geçersiz veri
          schema:
            type: string
        "500":
          description: Todo eklenemedi
          schema:
            type: string
      summary: Create a new Todo
      tags:
      - todos
  /todos/{id}:
    delete:
      description: Bu endpoint, belirli bir todo'yu siler.
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Todo başarıyla silindi
          schema:
            type: string
        "400":
          description: Geçersiz ID
          schema:
            type: string
        "500":
          description: Todo silinemedi
          schema:
            type: string
      summary: Delete a Todo
      tags:
      - todos
    get:
      description: Bu endpoint, belirli bir todo'yu getirir.
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Belirli todo başarıyla getirildi
          schema:
            $ref: '#/definitions/ent.Todo'
        "400":
          description: Geçersiz ID
          schema:
            type: string
        "404":
          description: Todo bulunamadı
          schema:
            type: string
        "500":
          description: Veritabanı hatası
          schema:
            type: string
      summary: Get a Todo by ID
      tags:
      - todos
    put:
      consumes:
      - application/json
      description: Bu endpoint, belirli bir todo'yu günceller.
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: Todo nesnesi
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/ent.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: Todo başarıyla güncellendi
          schema:
            type: string
        "400":
          description: Geçersiz veri
          schema:
            type: string
        "500":
          description: Todo güncellenemedi
          schema:
            type: string
      summary: Update a Todo
      tags:
      - todos
  /users/login:
    post:
      consumes:
      - application/json
      description: Bu endpoint, kullanıcı adı ve şifre ile giriş yapmayı sağlar.
      parameters:
      - description: Kullanıcı Giriş Bilgileri
        in: body
        name: loginUser
        required: true
        schema:
          properties:
            password:
              type: string
            username:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Başarıyla giriş yapıldı
          schema:
            $ref: '#/definitions/ent.User'
        "401":
          description: Yetkilendirme başarısız
          schema:
            type: string
        "500":
          description: Sunucu hatası
          schema:
            type: string
      summary: Login
      tags:
      - Kullanıcı
swagger: "2.0"